64ビット専用にしたのとファイル周りのエラーチェックを変更

+++ src/Models/Apery.cs
@@ -10,15 +10,20 @@ using Amazon.Runtime;
 using Amazon.S3.Model;
 using Prism.Mvvm;
 using System.Threading.Tasks;
-using SevenZip.SDK.Compress.LZMA;
 using System.Reactive.Linq;
 
 namespace AperyGenerateTeacherGUI.Models
 {
     public sealed class Apery : BindableBase, IDisposable
     {
-        private string _log;
+        private Tuple<bool, bool> _isAperyIdleAndStarted = new Tuple<bool, bool>(false,false);
+        public Tuple<bool, bool> IsAperyIdleAndStarted 
+        {
+            get { return this._isAperyIdleAndStarted; }
+            set { this.SetProperty(ref this._isAperyIdleAndStarted, value); }
+        }
 
+        private string _log;
         public string Log
         {
             get { return this._log; }
@@ -26,7 +31,6 @@ namespace AperyGenerateTeacherGUI.Models
         }
 
         private bool _isAperyIdle;
-
         public bool IsAperyIdle
         {
             get { return this._isAperyIdle; }
@@ -34,19 +38,17 @@ namespace AperyGenerateTeacherGUI.Models
         }
 
         private double _progress;
-
         public double Progress
         {
             get { return this._progress; }
             set { this.SetProperty(ref this._progress, value); }
         }
-
         private string _outFile;
         private readonly Random _random;
         private readonly Process _aperyInstance;
         private readonly IDisposable _aperyStandardOutput;
         private static readonly Lazy<Apery> _apery = new Lazy<Apery>(() => new Apery());
-
+        private bool _isAperyStarted;
         public static Apery Instance => _apery.Value;
 
         private Apery()
@@ -59,9 +61,8 @@ namespace AperyGenerateTeacherGUI.Models
                 UseShellExecute = false
             };
 
-            this._aperyInstance = new Process() { StartInfo = startInfo };
-            this._aperyInstance.Start();
-            this.IsAperyIdle = true;
+            this._aperyInstance = new Process { StartInfo = startInfo };
+
             this._random = new Random();
 
             this._aperyStandardOutput = Observable.FromEvent<DataReceivedEventHandler, DataReceivedEventArgs>(
@@ -88,7 +89,17 @@ namespace AperyGenerateTeacherGUI.Models
                          this._aperyInstance.CancelOutputRead();
                          TreatResultAsync();
                      }
-                 });
+                 }
+                 );
+        }
+
+        public void Run(bool canRun)
+        {
+            if (!canRun) return;
+            this._aperyInstance.Start();
+            this.IsAperyIdleAndStarted = new Tuple<bool, bool>(true,true);
+            this.IsAperyIdle = true;
+            this._isAperyStarted = true;
         }
 
         private async void TreatResultAsync()
@@ -109,7 +120,7 @@ namespace AperyGenerateTeacherGUI.Models
                     UseShellExecute = false
                 };
 
-                using (var process = new Process() { StartInfo = shuffleStartInfo })
+                using (var process = new Process { StartInfo = shuffleStartInfo })
                 {
                     process.Start();
                     process.WaitForExit(); //本家v1.0.1から
@@ -121,56 +132,14 @@ namespace AperyGenerateTeacherGUI.Models
                 #endregion
             });
 
-//            #region 教師データ圧縮
-//            this.Log = "教師データ圧縮中";
-
-//            var outCompressedFile = $"{shufOutfile}.7z";
-
-//            if (await CompressFileAsync(shufOutfile, outCompressedFile))
-//            {
-//                this.Log = "教師データ圧縮完了";
-//                File.Delete(shufOutfile);
-
-//                #endregion
-
-//                #region send aws s3
-//#if DEBUG
-//    /*
-//#endif
-//                await SendResultToAwsAsync(outCompressedFile);
-//#if DEBUG
-//                */
-//#endif
-//                #endregion
-
-//            }
-//            else
-//            {
-//                this.Log = "教師データの圧縮に失敗しました";
-//                File.Delete(shufOutfile);
-//            }
-
             await SendResultToAwsAsync(shufOutfile);
-            File.Delete(shufOutfile);
 
             this.IsAperyIdle = true;
         }
 
-
-        public async void RunProcessAsync(short threads, long teacherNodes)
-        {
-            await Task.Run(() =>
-            {
-                this.IsAperyIdle = false;
-                this._outFile = $"out_{RandomString(20)}.hcpe";
-                this._aperyInstance.StandardInput.WriteLine($"make_teacher roots.hcp {this._outFile} {threads} {teacherNodes}");
-                this._aperyInstance.BeginOutputReadLine();
-            })
-            ;
-        }
-
         public void RunProcess(short threads, long teacherNodes)
         {
+            this.IsAperyIdleAndStarted=new Tuple<bool, bool>(false,true);
             this.IsAperyIdle = false;
             this._outFile = $"out_{RandomString(20)}.hcpe";
             this._aperyInstance.StandardInput.WriteLine($"make_teacher roots.hcp {this._outFile} {threads} {teacherNodes}");
@@ -180,6 +149,7 @@ namespace AperyGenerateTeacherGUI.Models
 
         private async Task SendResultToAwsAsync(string filePath)
         {
+            this.Log = "sending...";
             try
             {
                 using (var amazonS3Client = new AmazonS3Client(new AnonymousAWSCredentials(), RegionEndpoint.USWest1))
@@ -195,14 +165,7 @@ namespace AperyGenerateTeacherGUI.Models
 
                         var response = await amazonS3Client.PutObjectAsync(request);
 
-                        if (response.HttpStatusCode == HttpStatusCode.Accepted)
-                        {
-                            this.Log = "サーバーに教師データを送信完了しました。ご協力ありがとうございました。";
-                        }
-                        else
-                        {
-                            this.Log = "サーバーに何か問題あるかも？";
-                        }
+                        this.Log = response.HttpStatusCode == HttpStatusCode.OK ? "サーバーに教師データを送信完了しました。ご協力ありがとうございました。" : $"サーバーに何か問題あるかも？ : {response.HttpStatusCode.ToString()}";
                     }
                     catch (AmazonS3Exception amazonS3Excetion)
                     {
@@ -231,105 +194,6 @@ namespace AperyGenerateTeacherGUI.Models
             }
 
         }
-
-
-        private void SendResultToAws(string filePath)
-        {
-            try
-            {
-                using (var amazonS3Client = new AmazonS3Client(new AnonymousAWSCredentials(), RegionEndpoint.USWest1))
-                {
-
-                    try
-                    {
-                        var request = new PutObjectRequest
-                        {
-                            BucketName = "apery-teacher-v1.0.1",
-                            FilePath = filePath,
-                        };
-
-
-                        var response =
-                            amazonS3Client
-                            .PutObject(request)
-                            ;
-
-                        this.Log = "サーバーに教師データを送信完了しました。ご協力ありがとうございました。";
-
-                    }
-                    catch (AmazonS3Exception amazonS3Excetion)
-                    {
-                        if (amazonS3Excetion.ErrorCode != null &&
-                            (amazonS3Excetion.ErrorCode.Equals("InvalidAccessKeyId") ||
-                             amazonS3Excetion.ErrorCode.Equals("InvalidSeculity")))
-                        {
-                            this.Log = "サーバーにアクセス出来ませんでした。";
-                        }
-                        else
-                        {
-                            this.Log = "サーバーへの教師データデータ送信に失敗しました。";
-                        }
-                    }
-
-
-                }
-            }
-            catch (Exception)
-            {
-                this.Log = "サーバー接続に失敗しました。";
-            }
-            finally
-            {
-                //ファイル送信のループ処理がないなら削除するタイミングはここでは
-                File.Delete(filePath);
-            }
-        }
-
-
-
-        private static void CompressFile(string inFile, string outFile)
-        {
-            var coder = new Encoder();
-
-            using (var input = new FileStream(inFile, FileMode.Open))
-            {
-                using (var output = new FileStream(outFile, FileMode.Create))
-                {
-                    coder.WriteCoderProperties(output);
-                    output.Write(BitConverter.GetBytes(input.Length), 0, 8);
-
-                    coder.Code(input, output, input.Length, -1, null);
-                    output.Flush();
-                }
-            }
-        }
-
-        private static async Task<bool> CompressFileAsync(string inFile, string outFile)
-        {
-            try
-            {
-                var coder = new Encoder();
-
-                using (var input = new FileStream(inFile, FileMode.Open))
-                {
-                    using (var output = new FileStream(outFile, FileMode.Create))
-                    {
-                        coder.WriteCoderProperties(output);
-                        output.Write(BitConverter.GetBytes(input.Length), 0, 8);
-
-                        coder.Code(input, output, input.Length, -1, null);
-                        await output.FlushAsync();
-
-                        return true;
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-        }
-
 
         #region IDisposable Support
         private bool disposedValue = false; // To detect redundant calls
@@ -341,10 +205,14 @@ namespace AperyGenerateTeacherGUI.Models
                 if (disposing)
                 {
                     // TODO: dispose managed state (managed objects).
-                    if (!this._aperyInstance.HasExited)
+                    if (this._isAperyStarted)
                     {
-                        this._aperyInstance.Kill();
+                        if (!this._aperyInstance.HasExited)
+                        {
+                            this._aperyInstance.Kill();
+                        }
                     }
+
                     this._aperyInstance.Dispose();
 
                     this._aperyStandardOutput.Dispose();
@@ -384,10 +252,7 @@ namespace AperyGenerateTeacherGUI.Models
                 .Select(_ => candidates[this._random.Next(candidatesLength)])
                 .ToArray()
                 );
-
         }
-
-
 
     }
 }
